+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/AuthnResponseResult.kt
@@ -1,6 +1,7 @@
 package at.asitplus.wallet.lib.openid
 
 import at.asitplus.openid.dcql.DCQLCredentialQueryIdentifier
+import at.asitplus.wallet.lib.agent.validation.CredentialFreshnessSummary
 import at.asitplus.wallet.lib.agent.validation.CredentialTimelinessValidationSummary
 import at.asitplus.wallet.lib.data.IsoDocumentParsed
 import at.asitplus.wallet.lib.data.SelectiveDisclosureItem
@@ -53,9 +54,16 @@ sealed class AuthnResponseResult {
         val reconstructed: JsonObject,
         val disclosures: Collection<SelectiveDisclosureItem>,
         val state: String?,
-        val timelinessValidationSummary: CredentialTimelinessValidationSummary.SdJwt,
-        val tokenStatus: TokenStatusValidationResult,
-    ) : AuthnResponseResult()
+        val freshnessSummary: CredentialFreshnessSummary.SdJwt,
+    ) : AuthnResponseResult() {
+        @Deprecated("Replaced with more expressive TokenStatusValidationResult, supporting token status values as defined by the library client.", ReplaceWith("freshnessSummary.tokenStatusValidationResult"))
+        val tokenStatus: TokenStatusValidationResult
+            get() = freshnessSummary.tokenStatusValidationResult
+
+        @Deprecated("", ReplaceWith("credentialFreshnessSummary.timelinessValidationSummary"))
+        val timelinessValidationSummary: CredentialTimelinessValidationSummary.SdJwt
+            get() = freshnessSummary.timelinessValidationSummary
+    }
 
     /**
      * Successfully decoded and validated the response from the Wallet (ISO credential)