+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/PreRegisteredClientTest.kt
@@ -85,7 +85,7 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
+        result.vp.freshVerifiableCredentials.shouldNotBeEmpty()
 
         verifierOid4vp.createAuthnRequest(
             defaultRequestOptions, OpenId4VpVerifier.CreationOptions.Query(walletUrl)
@@ -116,7 +116,7 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
+        result.vp.freshVerifiableCredentials.shouldNotBeEmpty()
         result.state.shouldBe(expectedState)
     }
 
@@ -206,7 +206,7 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.params.formUrlEncode())
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
+        result.vp.freshVerifiableCredentials.shouldNotBeEmpty()
     }
 
     "test with direct_post_jwt" {
@@ -231,7 +231,7 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.params.formUrlEncode())
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
+        result.vp.freshVerifiableCredentials.shouldNotBeEmpty()
     }
 
     "test with deserializing" {
@@ -250,7 +250,7 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponseParams)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
+        result.vp.freshVerifiableCredentials.shouldNotBeEmpty()
     }
 
     "test specific credential" {
@@ -264,8 +264,8 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
-        result.vp.timelyVerifiableCredentials.map { it.vcJws }.forEach {
+        result.vp.freshVerifiableCredentials.shouldNotBeEmpty()
+        result.vp.freshVerifiableCredentials.map { it.vcJws }.forEach {
             it.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
         }
     }
@@ -280,8 +280,8 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
-        result.vp.timelyVerifiableCredentials.map { it.vcJws }.forEach {
+        result.vp.freshVerifiableCredentials.shouldNotBeEmpty()
+        result.vp.freshVerifiableCredentials.map { it.vcJws }.forEach {
             it.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
         }
     }
@@ -312,8 +312,8 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
-        result.vp.timelyVerifiableCredentials.map { it.vcJws }.forEach {
+        result.vp.freshVerifiableCredentials.shouldNotBeEmpty()
+        result.vp.freshVerifiableCredentials.map { it.vcJws }.forEach {
             it.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
         }
     }
@@ -338,8 +338,8 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
-        result.vp.timelyVerifiableCredentials.map { it.vcJws }.forEach {
+        result.vp.freshVerifiableCredentials.shouldNotBeEmpty()
+        result.vp.freshVerifiableCredentials.map { it.vcJws }.forEach {
             it.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
         }
     }