+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/validation/TimeScope.kt
@@ -1,6 +1,5 @@
 package at.asitplus.wallet.lib.agent.validation
 
-import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
 import kotlin.time.Duration
 
@@ -14,12 +13,5 @@ data class TimeScope(
     fun Instant.isTooEarly() = this < earliestTime
     fun Instant.isTooLate() = this > latestTime
 
-    companion object {
-        fun <T> timeScoped(clock: Clock, timeLeeway: Duration, block: TimeScope.() -> T) = block(
-            TimeScope(
-                now = clock.now(),
-                timeLeeway = timeLeeway,
-            ),
-        )
-    }
+    operator fun <T> invoke(block: TimeScope.() -> T) = block(this)
 }