+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/validation/mdoc/MobileSecurityObjectTimelinessValidationSummary.kt
@@ -1,16 +1,18 @@
 package at.asitplus.wallet.lib.agent.validation.mdoc
 
+import at.asitplus.wallet.lib.agent.validation.TimelinessIndicator
 import at.asitplus.wallet.lib.agent.validation.common.EntityExpiredError
 import at.asitplus.wallet.lib.agent.validation.common.EntityNotYetValidError
 import kotlinx.datetime.Instant
 
 data class MobileSecurityObjectTimelinessValidationSummary(
-    val evaluationTime: Instant,
+    override val evaluationTime: Instant,
     val mdocExpiredError: EntityExpiredError?,
     val mdocNotYetValidError: EntityNotYetValidError?,
-) {
-    val isSuccess = listOf(
-        mdocExpiredError,
-        mdocNotYetValidError,
-    ).all { it == null }
+) : TimelinessIndicator {
+    override val isExpired: Boolean
+        get() = mdocExpiredError != null
+
+    override val isNotYetValid: Boolean
+        get() = mdocNotYetValidError != null
 }