+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/validation/mdoc/MobileSecurityObjectTimelinessValidator.kt
@@ -1,19 +1,16 @@
 package at.asitplus.wallet.lib.agent.validation.mdoc
 
-import at.asitplus.wallet.lib.agent.validation.TimeScope.Companion.timeScoped
+import at.asitplus.wallet.lib.agent.validation.TimeScope
 import at.asitplus.wallet.lib.agent.validation.common.EntityExpiredError
 import at.asitplus.wallet.lib.agent.validation.common.EntityNotYetValidError
 import at.asitplus.wallet.lib.iso.MobileSecurityObject
 import io.github.aakira.napier.Napier
-import kotlinx.datetime.Clock
-import kotlin.time.Duration
-import kotlin.time.Duration.Companion.seconds
 
-data class MobileSecurityObjectTimelinessValidator(
-    val timeLeeway: Duration = 300.seconds,
-    private val clock: Clock = Clock.System,
-) {
-    operator fun invoke(mobileSecurityObject: MobileSecurityObject) = timeScoped(clock, timeLeeway) {
+class MobileSecurityObjectTimelinessValidator {
+    operator fun invoke(
+        mobileSecurityObject: MobileSecurityObject,
+        timeScope: TimeScope,
+    ) = timeScope {
         MobileSecurityObjectTimelinessValidationSummary(
             evaluationTime = now,
             mdocExpiredError = if (mobileSecurityObject.validityInfo.validUntil.isTooEarly()) {
@@ -31,7 +28,7 @@ data class MobileSecurityObjectTimelinessValidator(
                 )
             } else null,
         ).also {
-            if (it.isSuccess) {
+            if (it.isTimely) {
                 Napier.d("MSO is timely")
             }
         }