+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/validation/vcJws/VcJwsTimelinessValidationDetails.kt
@@ -1,21 +1,21 @@
 package at.asitplus.wallet.lib.agent.validation.vcJws
 
+import at.asitplus.wallet.lib.agent.validation.TimelinessIndicator
 import at.asitplus.wallet.lib.agent.validation.common.EntityExpiredError
 import at.asitplus.wallet.lib.agent.validation.common.EntityNotYetValidError
 import kotlinx.datetime.Instant
 
 data class VcJwsTimelinessValidationDetails(
-    val evaluationTime: Instant,
+    override val evaluationTime: Instant,
     val jwsExpiredError: EntityExpiredError?,
     val credentialExpiredError: EntityExpiredError?,
     val jwsNotYetValidError: EntityNotYetValidError?,
     val credentialNotYetValidError: EntityNotYetValidError?,
-) {
-    val isSuccess = listOf(
-        jwsExpiredError,
-        credentialExpiredError,
-        jwsNotYetValidError,
-        credentialNotYetValidError,
-    ).all { it == null }
+): TimelinessIndicator {
+    override val isExpired: Boolean
+        get() = jwsExpiredError != null || credentialExpiredError != null
+
+    override val isNotYetValid: Boolean
+        get() = jwsNotYetValidError != null || credentialNotYetValidError != null
 }