+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/ValidatorVcTest.kt
@@ -6,10 +6,10 @@ import at.asitplus.signum.indispensable.josef.JwsHeader
 import at.asitplus.signum.indispensable.josef.JwsSigned
 import at.asitplus.signum.supreme.signature
 import at.asitplus.wallet.lib.agent.Verifier.VerifyCredentialResult
-import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatusValidationResult
 import at.asitplus.wallet.lib.data.*
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.StatusListInfo
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
+import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatusValidationResult
 import at.asitplus.wallet.lib.data.rfc3986.UniformResourceIdentifier
 import at.asitplus.wallet.lib.jws.JwsContentTypeConstants
 import at.asitplus.wallet.lib.jws.JwsHeaderKeyId
@@ -313,7 +313,7 @@ class ValidatorVcTest : FreeSpec() {
                         val validationResult = validator.verifyVcJws(it, verifierKeyMaterial.publicKey)
                             .shouldBeInstanceOf<VerifyCredentialResult.SuccessJwt>()
 
-                        validator.checkTimeliness(validationResult.jws).isSuccess shouldBe false
+                        validator.checkCredentialTimeliness(validationResult.jws).isTimely shouldBe false
                     }
             }
         }